requirements.txt
# Core trading dependencies (from ai-trading-agent)
hyperliquid-python-sdk>=0.20.0,<0.21.0
python-dotenv>=1.1.1,<2.0.0
web3>=7.14.0,<8.0.0
aiohttp>=3.13.1,<4.0.0
openai>=2.5.0,<3.0.0
requests>=2.32.5,<3.0.0

# GUI dependencies (NiceGUI desktop app)
nicegui>=2.0.0
plotly>=5.18.0
pandas>=2.1.0
pywebview>=5.0.0

# Database
sqlalchemy>=2.0.0

# Packaging
pyinstaller>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities

# Gemini SDK (for direct Gemini API integration)
google-generativeai>=0.8.0
